packages/core-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentid-core"
version = "0.1.0"
description = "AgentID core SDK - cryptographic identity, manifests, signing, verification, and policy evaluation for AI agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
keywords = ["ai", "agent", "identity", "verification", "cryptography", "ed25519", "manifest", "policy", "trust"]
dependencies = [
    "PyNaCl>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agentid"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
